requirements.txt
crawl4ai>=0.8.0
pydantic>=2.0
pandas>=2.0
snowflake-connector-python>=3.0
python-dotenv>=1.0
beautifulsoup4>=4.12.0
pytest>=7.0
pytest-asyncio>=0.21.0
